requirements.txt
# BetaCode - Dependências
# Python 3.11+ requerido

# YAML parsing
PyYAML>=6.0.1

# CLI
click>=8.1.7

# Type checking
typing-extensions>=4.9.0

# Testing (dev)
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality (dev)
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2

# Linting (dev)
pylint>=3.0.3

# Documentation (dev)
sphinx>=7.2.6
